# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yajja delivery & social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Vendors ───────────────────────────────────────────────────────────────────
...
  /groups:
    get:
      operationId: listMyGroups
      tags: [groups]
      summary: List groups I belong to
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/groupId"
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/groupId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
  /groups/{groupId}/members/{userId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove member from group (admin only)
      parameters:
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Messages (Group Chat) ─────────────────────────────────────────────────────
  /groups/{groupId}/messages:
    get:
      operationId: listGroupMessages
      tags: [messages]
      summary: Get group chat messages
      parameters:
        - $ref: "#/components/parameters/groupId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMessage"
    post:
      operationId: sendGroupMessage
      tags: [messages]
      summary: Send message in group chat
      parameters:
        - $ref: "#/components/parameters/groupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMessage"

  # ── Invites ───────────────────────────────────────────────────────────────────
  /invites:
    get:
      operationId: listMyInvites
      tags: [invites]
      summary: List pending invites for current user
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      operationId: sendInvite
      tags: [invites]
      summary: Send group invite to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInviteBody"
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /invites/{inviteId}/accept:
    post:
      operationId: acceptInvite
      tags: [invites]
      parameters:
        - $ref: "#/components/parameters/inviteId"
      responses:
        "200":
          description: Accepted, joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /invites/{inviteId}/decline:
    post:
      operationId: declineInvite
      tags: [invites]
      parameters:
        - $ref: "#/components/parameters/inviteId"
      responses:
        "200":
          description: Declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Leaderboard ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
...
    groupId:
      name: groupId
      in: path
      required: true
      schema:
        type: integer
...
    inviteId:
      name: inviteId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    UserRole:
      type: string
      enum: [customer, vendor, rider, admin]

    # ── Auth ────────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
        avatarUrl:
          type: string
        address:
          type: string
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]

    # ── Vendor ───────────────────────────────────────────────────────────────────
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        adminId:
          type: integer
        adminName:
          type: string
        imageUrl:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, adminId, memberCount, createdAt]
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
      required: [name]
    GroupMember:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        avatarUrl:
          type: string
        isAdmin:
          type: boolean
        joinedAt:
          type: string
      required: [id, groupId, userId, isAdmin, joinedAt]

    # ── Messages ──────────────────────────────────────────────────────────────────
    GroupMessage:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        avatarUrl:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, groupId, userId, content, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]

    # ── Invites ───────────────────────────────────────────────────────────────────
    Invite:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        groupName:
          type: string
        fromUserId:
          type: integer
        fromUserName:
          type: string
        toUserId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
      required: [id, groupId, fromUserId, toUserId, status, createdAt]
    SendInviteBody:
      type: object
      properties:
        groupId:
          type: integer
        toUserId:
          type: integer
      required: [groupId, toUserId]

    # ── Leaderboard ───────────────────────────────────────────────────────────────
...